Update: Champion

+++ src/service/ChampionService.ts
@@ -1,14 +1,17 @@
+import { ChampionInputError } from "../Error/Champion/ChampionError";
 import ChampionRepository from "../repository/ChampionRepository";
 import { ChampionProps } from "../types/Champion/Champion";
 import { UpdateChampionProps } from "../types/Champion/UpdateChampion";
-import { ChampionInputError } from "../Error/Champion/ChampionError";
 
 class ChampionService {
   async insertChampion(champion: ChampionProps) {
     const { name, price, role } = champion;
 
     if (!name || !price || !role) {
-      throw new ChampionInputError("Preencha todos os campos corretamente");
+      throw new ChampionInputError(
+        "Preencha todos os campos corretamente",
+        400
+      );
     }
 
     const existDoc = await ChampionRepository.showByName(name);
@@ -16,44 +19,82 @@ class ChampionService {
       return ChampionRepository.store(champion);
     }
 
-    throw new ChampionInputError("Já existe um campeão com este nome.");
+    throw new ChampionInputError("Já existe um campeão com este nome.", 400);
   }
 
   async updateChampion(id: string, championDoc: UpdateChampionProps) {
-    const listOfChampion = await this.indexChampions();
+    const listOfChampion = await this.listChampions();
 
     const existingName = listOfChampion
       .map((item) => item.name)
       .find((champion) => champion === championDoc.name);
 
     if (existingName) {
-      throw new ChampionInputError("já tem");
+      throw new ChampionInputError("Já existe um campeão com este nome.", 400);
     }
     console.log("cheguei aqui");
     const updatedChampion = await ChampionRepository.update(id, championDoc);
     return updatedChampion;
   }
 
-  async indexChampions() {
+  async listChampions() {
     const listOfChampions = await ChampionRepository.index();
 
     if (listOfChampions.length === 0) {
-      throw new ChampionInputError("Nenhum campeão encontrado");
+      throw new ChampionInputError("Nenhum campeão encontrado", 400);
     }
 
     return listOfChampions;
   }
 
-  showOneChampionById(id: string) {
-    return ChampionRepository.show(id);
+  async listChampionsDoc() {
+    const listOfChampions = await ChampionRepository.indexAll();
+
+    if (listOfChampions.length === 0) {
+      throw new ChampionInputError("Nenhum campeão encontrado", 400);
+    }
+
+    return listOfChampions;
   }
 
-  showOneChampionByName(name: string) {
-    return ChampionRepository.showByName(name);
+  async showOneChampionById(id: string) {
+    if (!id) {
+      throw new ChampionInputError("ID é um campo obrigatório", 400);
+    }
+
+    const champion = await ChampionRepository.show(id);
+
+    if (!champion) {
+      throw new ChampionInputError("ID de campeão não encontrado", 400);
+    }
+    return champion;
   }
 
-  destroyOneChampionById(id: string) {
-    return ChampionRepository.destroy(id);
+  async showOneChampionByName(name: string) {
+    if (!name) {
+      throw new ChampionInputError("Nome é um campo obrigatório", 400);
+    }
+
+    const champion = await ChampionRepository.showByName(name);
+
+    if (!champion) {
+      throw new ChampionInputError("Nome de campeão não encontrado", 400);
+    }
+    return champion;
+  }
+
+  async destroyOneChampionById(id: string) {
+    if (!id) {
+      throw new ChampionInputError("ID é um campo obrigatório", 400);
+    }
+
+    const champion = await ChampionRepository.destroy(id);
+
+    if (!champion) {
+      throw new ChampionInputError("ID de campeão não encontrado", 400);
+    }
+
+    return;
   }
 }
